requirements.txt
alembic==1.8.1
cachelib==0.9.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
Flask==2.2.2
Flask-Caching==2.0.1
Flask-Migrate==3.1.0
flask-paginate==2022.1.8
Flask-SQLAlchemy==3.0.2
greenlet==1.1.3.post0
idna==3.4
importlib-metadata==5.0.0
importlib-resources==5.10.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.3
MarkupSafe==2.1.1
psycopg2==2.9.5
python-dotenv==0.21.0
requests==2.28.1
SQLAlchemy==1.4.42
typing_extensions==4.4.0
urllib3==1.26.12
Werkzeug==2.2.2
zipp==3.10.0